Load design to behavioral graph builder.
